Extracted Blockquote to repo textbit-plugins. Moved util functions to TextbitEditor to be used in external plugins
