+++ local/src/plugins/List/lib/normalizeNode.tsx
@@ -1,9 +1,8 @@
 import { Editor, NodeEntry, Transforms, Node } from 'slate'
-import { TextbitElement } from '../../../../../src'
-import { hasText } from '../../../../../src/lib/utils'
+import { TextbitEditor, TextbitElement } from '../../../../../src'
 
 export const normalizeNode = (editor: Editor, nodeEntry: NodeEntry, listType: string) => {
-  const [node, path] = nodeEntry
+  const [, path] = nodeEntry
   const children = Array.from(Node.children(editor, path))
 
 
@@ -32,7 +31,7 @@ export const normalizeNode = (editor: Editor, nodeEntry: NodeEntry, listType: st
     // the remaining last node to normal text. This gives the appearance that
     // <enter> on a last empty list item converts it to a text node.
     if (n === children.length && children.length > 1 && TextbitElement.isOfType(child, `${listType}/list-item`)) {
-      if (!hasText([children[n - 2], children[n - 1]])) {
+      if (!TextbitEditor.hasText([children[n - 2], children[n - 1]])) {
 
         const removePath = [childPath[0], n - 2]
         const liftPath = [childPath[0], n - 1]
