+++ src/components/core/index.ts
@@ -1,6 +1,5 @@
 import { TextElement } from './TextElement'
 import { Bold, Italic, Underline } from "./Leafs"
-import { BlockquotePlugin } from "./Blockquote"
 import { Navigation } from "./Navigation"
 import { OEmbed } from "./OEmbed"
 import { Image } from "./Image"
@@ -11,7 +10,6 @@ import { Link } from './Link'
 
 export const StandardPlugins = [
   TextElement,
-  BlockquotePlugin,
   Bold,
   Italic,
   Underline,
